Pull Request for testing Refactor DeviceLib into a separate target.
The DeviceLib is to manage the real devices using usbmuxd: it monitors the device connection and set up socket connections. The ChannelLib uses this to manage the pool of connections to the device by its serial device UUID. This target can also be built used separately to connect to the real devices.

PiperOrigin-RevId: 235240676